refactor(dashboard): split the cockpit monolith — pure format lib + panel layer, semantic <time> stamps (cave-tsoz)
dashboard-cockpit.tsx was a ~1,200-line monolith: data fetching, derivation,
layout persistence, drag wiring, AND ~600 lines of presentational panels plus
a dozen pure helpers. Slice it along its natural seams (the chat-familiar-view
extraction from #3149 is the pattern):

- src/lib/dashboard-cockpit-format.ts — pure, DOM-free helpers + vocab:
  layout reconciliation (Layout/DEFAULT_LAYOUT/PANEL_TITLES/panelTitle/
  reconcileLayout), trend store (TrendKey/DaySnap/TrendStore/dayKey/
  seriesFor), status/health/tier vocab, KPI sub-lines (wowSub/retroSub/
  contractSub/coverageSub), whenLabel, shortRepo, prettyFactor,
  confidenceColor, longDate — now behaviorally tested (layout reconcile,
  7-point series with honest gaps, teach-state sub-lines, calendar-relative
  labels, clamps).
- src/components/dashboard/cockpit-panels.tsx — the presentational layer
  (CovenInsightBanner, Panel, SortableWidget, KpiTile, FamiliarInsightsTable,
  Usage/Board/Agents/Github/Agenda/Signals/Confidence/SpaceUsage panels),
  each still owning its skeleton and empty state.
- dashboard-cockpit.tsx keeps what the root owns: fetching, derivation,
  trends persistence, drag orchestration, and the page frame (~600 lines).

Legibility: the three timestamp cells (agenda when-labels, insight rows'
last-active, space-usage Updated) are now semantic <time dateTime> elements.

Pins updated for the split: dashboard-page.test.ts reads the three files as
one contract; the polish test pins each concern where it now lives (and pins
the new <time> semantics); authed-image consumer list points at the panel
layer. New format test wired in run-tests.
