+++ src/components/dashboard/dashboard-cockpit-polish.test.ts
@@ -1,43 +1,66 @@
 // @ts-nocheck
 // Dashboard cockpit polish (cave-m4oq): empty KPI tiles teach instead of
 // shrugging, no fake flatlines under missing data, one-accent trend waves,
-// and a quiet insight note instead of a heavy banner.
+// and a quiet insight note instead of a heavy banner. The cockpit is split
+// across the root (data + layout), cockpit-panels.tsx (presentation), and
+// dashboard-cockpit-format.ts (pure labels) — cave-tsoz.
 import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 
 const cockpit = readFileSync(new URL("./dashboard-cockpit.tsx", import.meta.url), "utf8");
+const panels = readFileSync(new URL("./cockpit-panels.tsx", import.meta.url), "utf8");
+const format = readFileSync(new URL("../../lib/dashboard-cockpit-format.ts", import.meta.url), "utf8");
 const css = readFileSync(new URL("../../styles/dashboard.css", import.meta.url), "utf8");
 
 // ── Empty tiles teach the action that fills them ─────────────────────────────
-assert.match(cockpit, /fills in after the first retro run/, "Retro tile teaches when empty");
-assert.match(cockpit, /fills in once familiars have contracts/, "Contract tile teaches when empty");
+assert.match(format, /fills in after the first retro run/, "Retro tile teaches when empty");
+assert.match(format, /fills in once familiars have contracts/, "Contract tile teaches when empty");
 assert.match(cockpit, /fills in after growth reviews/, "Confidence tile teaches when empty");
 for (const shrug of ["no retro runs", "no contracts", "no scores yet"]) {
-  assert.doesNotMatch(cockpit, new RegExp(`"${shrug}"`), `dead-end sub "${shrug}" is gone`);
+  for (const [name, src] of [["cockpit", cockpit], ["panels", panels], ["format", format]]) {
+    assert.doesNotMatch(src, new RegExp(`"${shrug}"`), `dead-end sub "${shrug}" is gone from ${name}`);
+  }
 }
 
 // ── No fake flatline under a metric with no data ─────────────────────────────
 assert.match(
-  cockpit,
+  panels,
   /\{value == null \? null : \(\s*\n\s*<Sparkline points=\{series\} color="var\(--accent-presence\)" height=\{22\} \/>/,
   "KPI tiles skip the sparkline while the metric has no data, and use the shared accent",
 );
 
 // ── One accent for every trend wave ──────────────────────────────────────────
 assert.match(
-  cockpit,
+  panels,
   /<Sparkline points=\{r\.trend\} color="var\(--accent-presence\)" height=\{22\} \/>/,
   "Familiar-insight rows draw one-accent sparklines (identity color stays on the avatar)",
 );
 assert.match(
-  cockpit,
+  panels,
   /<Sparkline points=\{p\.trend\} color="var\(--accent-presence\)" height=\{20\} \/>/,
   "Agent-panel trends draw one-accent sparklines",
 );
 assert.doesNotMatch(
-  cockpit,
+  panels,
   /<Sparkline[^>]*color=\{r\.color\}/,
   "No per-row rainbow sparkline remains",
+);
+
+// ── Timestamps are semantic <time> elements (cave-tsoz) ──────────────────────
+assert.match(
+  panels,
+  /<time className="cockpit-agendarow__when" dateTime=\{i\.fireAt!\}>/,
+  "agenda reminders carry machine-readable datetimes",
+);
+assert.match(
+  panels,
+  /<time dateTime=\{r\.lastActiveAt\}>\{relativeTime\(r\.lastActiveAt\)/,
+  "insight rows' last-active is a semantic time element",
+);
+assert.match(
+  panels,
+  /<time dateTime=\{new Date\(r\.lastModifiedMs\)\.toISOString\(\)\}>/,
+  "space-usage Updated column is a semantic time element",
 );
 
 // ── Insight banner reads as a quiet note ─────────────────────────────────────
